refactor file names and class names in vision examples
