+++ flygym/examples/vision/__init__.py
@@ -1,7 +1,7 @@
-from .arena import MovingFlyArena, MovingBarArena, ObstacleOdorArena
+from .arena import MovingObjArena, MovingFlyArena, MovingBarArena, ObstacleOdorArena
 from .vision_network import (
     RealTimeVisionNetwork,
     RealTimeVisionNetworkView,
     RetinaMapper,
 )
-from .fly_following_controller import NMFRealisticVision
+from .realisitc_vision import NMFRealisticVision
