--- flygym/examples/vision/obstacle_arena.py
@@ -1,165 +1,0 @@
-import numpy as np
-from typing import Union, Tuple, List, Optional, Callable
-
-from flygym.arena import BaseArena
-
-
-class ObstacleOdorArena(BaseArena):
-    num_sensors = 4
-
-    def __init__(
-        self,
-        terrain: BaseArena,
-        obstacle_positions: np.ndarray = np.array([(7.5, 0), (12.5, 5), (17.5, -5)]),
-        obstacle_colors: Union[np.ndarray, Tuple] = (0, 0, 0, 1),
-        obstacle_radius: float = 1,
-        obstacle_height: float = 4,
-        odor_source: np.ndarray = np.array([[25, 0, 2]]),
-        peak_odor_intensity: np.ndarray = np.array([[1]]),
-        diffuse_func: Callable = lambda x: x**-2,
-        marker_colors: Optional[List[Tuple[float, float, float, float]]] = None,
-        marker_size: float = 0.1,
-        user_camera_settings: Optional[
-            Tuple[Tuple[float, float, float], Tuple[float, float, float], float]
-        ] = None,
-    ):
-        self.terrain_arena = terrain
-        self.obstacle_positions = obstacle_positions
-        self.root_element = terrain.root_element
-        self.friction = terrain.friction
-        self.obstacle_radius = obstacle_radius
-        z_offset = terrain.get_spawn_position(np.zeros(3), np.zeros(3))[0][2]
-        obstacle_colors = np.array(obstacle_colors)
-        if obstacle_colors.shape == (4,):
-            obstacle_colors = np.array(
-                [obstacle_colors for _ in range(obstacle_positions.shape[0])]
-            )
-        else:
-            assert obstacle_colors.shape == (obstacle_positions.shape[0], 4)
-
-        self.odor_source = np.array(odor_source)
-        self.peak_odor_intensity = np.array(peak_odor_intensity)
-        self.num_odor_sources = self.odor_source.shape[0]
-        if self.odor_source.shape[0] != self.peak_odor_intensity.shape[0]:
-            raise ValueError(
-                "Number of odor source locations and peak intensities must match."
-            )
-        self.odor_dim = self.peak_odor_intensity.shape[1]
-        self.diffuse_func = diffuse_func
-
-        # Add markers at the odor sources
-        if marker_colors is None:
-            rgb = np.array([255, 127, 14]) / 255
-            marker_colors = [(*rgb, 1)] * self.num_odor_sources
-            num_odor_sources = self.odor_source.shape[0]
-        self.marker_colors = marker_colors
-        self._odor_marker_geoms = []
-        for i, (pos, rgba) in enumerate(zip(self.odor_source, marker_colors)):
-            pos = list(pos)
-            pos[2] += z_offset
-            marker_body = self.root_element.worldbody.add(
-                "body", name=f"odor_source_marker_{i}", pos=pos, mocap=True
-            )
-            geom = marker_body.add(
-                "geom", type="capsule", size=(marker_size, marker_size), rgba=rgba
-            )
-            self._odor_marker_geoms.append(geom)
-
-        # Reshape odor source and peak intensity arrays to simplify future calculations
-        _odor_source_repeated = self.odor_source[:, np.newaxis, np.newaxis, :]
-        _odor_source_repeated = np.repeat(_odor_source_repeated, self.odor_dim, axis=1)
-        _odor_source_repeated = np.repeat(
-            _odor_source_repeated, self.num_sensors, axis=2
-        )
-        self._odor_source_repeated = _odor_source_repeated
-        _peak_intensity_repeated = self.peak_odor_intensity[:, :, np.newaxis]
-        _peak_intensity_repeated = np.repeat(
-            _peak_intensity_repeated, self.num_sensors, axis=2
-        )
-        self._peak_intensity_repeated = _peak_intensity_repeated
-
-        # Add obstacles
-        self.obstacle_bodies = []
-        obstacle_material = self.root_element.asset.add(
-            "material", name="obstacle", reflectance=0.1
-        )
-        self.obstacle_z_pos = z_offset + obstacle_height / 2
-        for i in range(obstacle_positions.shape[0]):
-            obstacle_pos = [*obstacle_positions[i, :], self.obstacle_z_pos]
-            obstacle_color = obstacle_colors[i]
-            obstacle_body = self.root_element.worldbody.add(
-                "body", name=f"obstacle_{i}", mocap=True, pos=obstacle_pos
-            )
-            self.obstacle_bodies.append(obstacle_body)
-            obstacle_body.add(
-                "geom",
-                type="cylinder",
-                size=(obstacle_radius, obstacle_height / 2),
-                rgba=obstacle_color,
-                material=obstacle_material,
-            )
-
-        # Add monitor cameras
-        self.side_cam = self.root_element.worldbody.add(
-            "camera",
-            name="side_cam",
-            mode="fixed",
-            pos=(odor_source[0, 0] / 2, -25, 10),
-            euler=(np.deg2rad(75), 0, 0),
-            fovy=50,
-        )
-        self.back_cam = self.root_element.worldbody.add(
-            "camera",
-            name="back_cam",
-            mode="fixed",
-            pos=(-9, 0, 7),
-            euler=(np.deg2rad(60), 0, -np.deg2rad(90)),
-            fovy=55,
-        )
-        self.birdeye_cam = self.root_element.worldbody.add(
-            "camera",
-            name="birdeye_cam",
-            mode="fixed",
-            pos=(7.5, 0, 25),
-            euler=(0, 0, 0),
-            fovy=45,
-        )
-        self.birdeye_cam_origin = self.root_element.worldbody.add(
-            "camera",
-            name="birdeye_cam_origin",
-            mode="fixed",
-            pos=(0, 0, 40),
-            euler=(0, 0, 0),
-            fovy=50,
-        )
-        if user_camera_settings is not None:
-            cam_pos, cam_euler, cam_fovy = user_camera_settings
-            self.root_element.worldbody.add(
-                "camera",
-                name="user_cam",
-                mode="fixed",
-                pos=cam_pos,
-                euler=cam_euler,
-                fovy=cam_fovy,
-            )
-
-    def get_spawn_position(
-        self, rel_pos: np.ndarray, rel_angle: np.ndarray
-    ) -> Tuple[np.ndarray, np.ndarray]:
-        return self.terrain_arena.get_spawn_position(rel_pos, rel_angle)
-
-    def get_olfaction(self, antennae_pos: np.ndarray) -> np.ndarray:
-        antennae_pos_repeated = antennae_pos[np.newaxis, np.newaxis, :, :]
-        dist_3d = antennae_pos_repeated - self._odor_source_repeated  # (n, k, w, 3)
-        dist_euc = np.linalg.norm(dist_3d, axis=3)  # (n, k, w)
-        scaling = self.diffuse_func(dist_euc)  # (n, k, w)
-        intensity = self._peak_intensity_repeated * scaling  # (n, k, w)
-        return intensity.sum(axis=0)  # (k, w)
-
-    def pre_visual_render_hook(self, physics):
-        for geom, rgba in zip(self._odor_marker_geoms, self.marker_colors):
-            physics.bind(geom).rgba = np.array([*rgba[:3], 0])
-
-    def post_visual_render_hook(self, physics):
-        for geom, rgba in zip(self._odor_marker_geoms, self.marker_colors):
-            physics.bind(geom).rgba = np.array([*rgba[:3], 1])
